(svn r23316) -Feature: Add ability to zoom in to 2x and 4x level.

+++ src/spritecache.h
@@ -16,11 +16,11 @@
 
 /** Data structure describing a sprite. */
 struct Sprite {
-	byte height;  ///< Height of the sprite.
-	uint16 width; ///< Width of the sprite.
-	int16 x_offs; ///< Number of pixels to shift the sprite to the right.
-	int16 y_offs; ///< Number of pixels to shift the sprite downwards.
-	byte data[];  ///< Sprite data.
+	uint16 height; ///< Height of the sprite.
+	uint16 width;  ///< Width of the sprite.
+	int16 x_offs;  ///< Number of pixels to shift the sprite to the right.
+	int16 y_offs;  ///< Number of pixels to shift the sprite downwards.
+	byte data[];   ///< Sprite data.
 };
 
 extern uint _sprite_cache_size;
